Move Signed in its own module.
